requirements.txt
# Core dependencies
aiomqtt==2.4.0
pyyaml